### 0.0.12 Release * Issue #52 download vaccination timeline data file * Issue #53 import vaccination timeline data file into db * Issue #54 /vaccination/imported * Issue #55 /vaccination/tasks * Issue #56 /vaccination/info

+++ org/woehlke/covid19/europe/europe_service_download.py
@@ -29,7 +29,8 @@ class EuropeServiceDownload:
         os.makedirs('data', exist_ok=True)
         app.logger.info("------------------------------------------------------------")
         try:
-            os.remove(self.__src_cvsfile_name)
+            if os.path.isfile(self.__src_cvsfile_name):
+                os.remove(self.__src_cvsfile_name)
             wget.download(self.__url_src_data, self.__src_cvsfile_name)
             app.logger.info("------------------------------------------------------------")
         except RuntimeError as error:
