+++ org/woehlke/covid19/vaccination/vaccination_service_download.py
@@ -28,7 +28,8 @@ class VaccinationServiceDownload:
         app.logger.info("------------------------------------------------------------")
         os.makedirs('data', exist_ok=True)
         try:
-            os.remove(self.__src_cvsfile_name)
+            if os.path.isfile(self.__src_cvsfile_name):
+                os.remove(self.__src_cvsfile_name)
             data_file = wget.download(self.__url_src_data, self.__src_cvsfile_name)
             app.logger.info(" " + data_file + " ")
         except RuntimeError as error:
